Relocate components file
